+++ tools/highlight/tool.js
@@ -1,7 +1,7 @@
 (()=>{
   class HighlightTool extends LiChessTools.Tools.ToolBase {
 
-    dependencies=['EmitRedraw','DetectThirdParties'];
+    dependencies=['EmitRedraw','DetectThirdParties','TranspositionBehavior'];
 
     preferences=[
       {