+++ tools/playerLag/tool.js
@@ -44,6 +44,7 @@
     };
 
     _lagCache=new Map();
+    _useUserApi=true;
     getLag=async (username)=>{
       const parent=this.lichessTools;
       const $=parent.$;
@@ -51,8 +52,14 @@
       if (item && Date.now()-item.time<=this.opponentLagFrequency) {
         return item.value;
       }
-      const html=await parent.net.fetch({url:'/@/{username}/mini',args:{username}});
-      const lagRating=$(html).find('signal')[0]?.className.substr(1);
+      let lagRating;
+      if (this._useUserApi) {
+        const data=await parent.net.json({url:'/api/users/status?ids={username}&withSignal=true',args:{username}});
+        lagRating=data[0]?.signal;
+      } else {
+        const html=await parent.net.fetch({url:'/@/{username}/mini',args:{username}});
+        lagRating=$(html).find('signal')[0]?.className.substr(1);
+      }
       const lag=[750,500,300,150,75][lagRating];
       item={ time:Date.now(), value:lag };
       this._lagCache.set(username,item);
@@ -177,7 +184,7 @@
       };
       lichess.pubsub.off('socket.lag',this.onLag);
       lichess.pubsub.off('socket.in.mlat',this.onLatency);
-      $('.round__app .ruser-top a.user-link canvas.lichessTools-playerLag,.round__app .ruser-bottom a.user-link canvas.lichessTools-playerLag').remove();
+      $('.round__app .ruser-top a.user-link .lichessTools-playerLag,.round__app .ruser-bottom a.user-link .lichessTools-playerLag').remove();
       if (!this.options.isSet) return;
       lichess.pubsub.emit('socket.send', 'moveLat', true);
       lichess.pubsub.on('socket.lag',this.onLag);